Refactor RunningModule into its own file.
Also, clean up some requires.
